+++ CHANGELOG.md
@@ -15,6 +15,7 @@ method.
 
 ### Patches
 * Add version gating to some tests introduced in 1.5.0 [PR #128](https://github.com/corretto/amazon-corretto-crypto-provider/pull/128)
+* Correctly reject non-empty `PSource.PSpecified` values for RSA-OAEP.
 
 ## 1.6.0 (Never Released)
 This entry is for the planned, but never released, version 1.6.0 to assist in cross-references from the release notes in 1.5.0.